add: init game and player

+++ inc/cub3d.h
@@ -25,14 +25,12 @@
 # include "game.h"
 # include "debug.h"
 
-# include "../mlx_linux/mlx.h"
-# include "../mlx_linux/mlx_int.h"
-
+# include "mlx.h"
 
 typedef struct s_pos
 {
-    int x;
-    int y;
+    float x;
+    float y;
     char pos;
 } t_pos;
